Rename wizard-import-local -> wizard-import-choose-source
