+++ kolibri/plugins/management/assets/test/views/manage-content-page.spec.js
@@ -3,7 +3,6 @@ const Vue = require('vue-test');
 const Vuex = require('vuex');
 const simulant = require('simulant');
 const mutations = require('../../src/state/mutations');
-const sinon = require('sinon');
 const assert = require('assert');
 const ManageContentPage = require('../../src/views/manage-content-page/index.vue');
 
@@ -12,9 +11,6 @@ Vue.use(Vuex);
 function makeStore() {
   return new Vuex.Store({
     mutations,
-    actions: {
-      cancelImportExportWizard() {},
-    },
     state: {
       core: {
         channels: {
@@ -35,9 +31,9 @@ function makeStore() {
 function makeVm(options = {}) {
   const store = options.store || makeStore();
   const components = {
-    'channels-grid': '<div>Channel Grid</div>',
+    'channels-grid': '<div>Channels Grid</div>',
     'task-status': '<div>Task Status</div>',
-    'wizard-import-local': '<div>Wizard</div>',
+    'wizard-import-choose-source': '<div>Wizard</div>',
   };
   const vuex = {
     actions: {
@@ -48,16 +44,15 @@ function makeVm(options = {}) {
   return new Ctor(Object.assign(options, { components, store, vuex })).$mount();
 }
 
-describe.only('manage content page index', () => {
-  it('clicking "import" goes to "wizard-import-local"', () => {
-    // wizard-import-local needs to be renamed
+describe('manage content page index', () => {
+  it('clicking "import" goes to "wizard-import-choose-source"', () => {
     const vm = makeVm();
     const importButton = vm.$el.querySelector('button[name="import"]');
     simulant.fire(importButton, 'click');
 
     return Vue.nextTick()
     .then(() => {
-      assert.equal(vm.$store.state.pageState.wizardState.page, 'CHOOSE_IMPORT_SOURCE');
+      assert.equal(vm.wizardComponent, 'wizard-import-choose-source');
     });
   });
 });
